requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.30.5
httpx==0.27.0
pinecone-client==3.0.2
tqdm==4.66.1
python-dotenv==1.0.1
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
tiktoken==0.5.1
pydantic==2.6.1
gunicorn==21.2.0